refactor: 유연한 사용을 위한 Query 인터페이스를 클래스로 변경
